ft_atoi

+++ ft_printf/libft/libft.h
@@ -63,4 +63,4 @@ void	ft_lstdelone(t_list *lst, void (*del)(void *));
 void	ft_lstclear(t_list **lst, void (*del)(void *));
 void	ft_lstiter(t_list *lst, void (*f)(void *));
 t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
-ssize_t	ft_atoi(const char *nptr);
+int		ft_atoi(const char *nptr);
